Dedicated module for ol.proj.Projection
